requirements.txt
click==8.1.3
ffmpeg-python==0.2.0
Flask==2.1.3
Flask-BasicAuth==0.2.0
Flask-WTF==1.0.1
future==0.18.2
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.1
WTForms==3.0.1
zipp==3.8.1
